test(distribution): refactored tests into separate files
